ProfilerLauncherPage shifted to MVVM design.
