+++ tools/Profiler/NanoProfiler/Helpers/Events.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace nanoFramework.Tools.NanoProfiler.Helpers
+{
+    public delegate void NotifyDelegate<in T>(T t);
+}
