+++ tools/Profiler/NanoProfiler/Helpers/ProfilingState.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace nanoFramework.Tools.NanoProfiler.Helpers
+{
+    public enum ProfilingState
+    {
+        Disconnected,
+        Connecting,
+        Connected,
+    }
+}
