+++ tools/Profiler/NanoProfiler/Views/ProfilerLauncherView.xaml
@@ -1,0 +1,50 @@
+﻿<Window x:Class="nanoFramework.Tools.NanoProfiler.Views.ProfilerLauncherView"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:nanoFramework.Tools.NanoProfiler.Views"
+        xmlns:viewModel="clr-namespace:nanoFramework.Tools.NanoProfiler.ViewModels"
+        mc:Ignorable="d"
+        Title="Profiler Launcher" Height="450" Width="800" 
+        Unloaded="Window_Unloaded" Loaded="Window_Loaded">
+    <Window.DataContext>
+        <viewModel:ProfileLauncherViewModel/>
+    </Window.DataContext>
+    <Grid Background="{Binding BackgroundProfileLauncher}">
+        <Label Content="nanoDevice to connect to:" HorizontalAlignment="Left" Margin="22,19,0,0" VerticalAlignment="Top"/>
+        <TextBox HorizontalAlignment="Left" Margin="22,50,0,0" TextWrapping="Wrap" Text="COM1" VerticalAlignment="Top" Width="120"/>
+        <Button x:Name="ConnectButton" IsEnabled="{Binding ConnectButtonEnabled}"
+                Content="{Binding ConnectButtonContent}" HorizontalAlignment="Left" Margin="22,84,0,0" VerticalAlignment="Top" Width="120" 
+                Command="{Binding ConnectClickedCommand}"/>
+
+        <GroupBox Header="Profiling Options" Margin="224,19,295,273">
+            <Grid Height="107" Margin="0,0,47,0">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition Height="Auto"/>
+                </Grid.RowDefinitions>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="284*"/>
+                </Grid.ColumnDefinitions>
+                <CheckBox Grid.Column="0" Grid.Row="0" Content="Force Reboot on Connect" HorizontalAlignment="Left" Margin="10,10,0,0" 
+                          Name="checkReboot"  Height="23" IsChecked="{Binding RebootChecked}"/>
+                <CheckBox Grid.Column="0" Grid.Row="1" Content="Profile Function Calls" HorizontalAlignment="Left" Margin="10,0,0,0" 
+                          Name="checkCalls" IsChecked="{Binding CallsChecked}" Height="23"/>
+                <CheckBox Grid.Column="0" Grid.Row="2" Content="Profile Object Allocations" HorizontalAlignment="Left" Margin="10,0,0,0" 
+                          Name="checkAllocations" IsChecked="{Binding AllocationsChecked}" Height="23"/>
+                <CheckBox Grid.Column="0" Grid.Row="3" Content="User Heap Absolute Address" HorizontalAlignment="Left" Margin="10,0,0,0" 
+                          Name="HeapAbsoluteAddress" Height="23" IsChecked="{Binding AbsoluteAddressChecked}"/>
+            </Grid>
+        </GroupBox>
+        <TextBox x:Name="textLog" Text="{Binding TextBlockContent}" 
+                 HorizontalAlignment="Left" Margin="22,197,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="574" Height="152"/>
+        <Button Content="Clear" HorizontalAlignment="Left" Margin="22,366,0,0" VerticalAlignment="Top" Width="91" 
+                Command="{Binding ClearClickedCommand}"/>
+        <Button x:Name="OpenLogFile" Content="Open Log File" HorizontalAlignment="Left" 
+                Margin="638,35,0,0" VerticalAlignment="Top" Width="103" Command="{Binding OpenLogFileCommand}" />
+
+    </Grid>
+</Window>
